[#230] feat: 목표 카드 UI 구성
