+++ lib/presentation/my_page/my_page.dart
@@ -1,7 +1,7 @@
 export 'providers/add_resolution_provider.dart';
 export 'providers/my_page_resolution_list_provider.dart';
 export 'providers/providers.dart';
-export 'screens/add_resolution_screen.dart';
-export 'screens/my_page_view.dart';
-export 'screens/my_page_view_widget.dart';
-export 'screens/screens.dart';
+export 'view/add_resolution_screen.dart';
+export 'view/my_page_view.dart';
+export 'view/my_page_view_widget.dart';
+export 'view/view.dart';
