+++ lib/presentation/my_page/view/add_resolution_screen.dart
