+++ lib/presentation/presentation.dart
@@ -67,9 +67,9 @@ export 'my_page/my_page.dart';
 export 'my_page/providers/add_resolution_provider.dart';
 export 'my_page/providers/my_page_resolution_list_provider.dart';
 export 'my_page/providers/providers.dart';
-export 'my_page/screens/add_resolution_screen.dart';
-export 'my_page/screens/my_page_view.dart';
-export 'my_page/screens/screens.dart';
+export 'my_page/view/add_resolution_screen.dart';
+export 'my_page/view/my_page_view.dart';
+export 'my_page/view/view.dart';
 export 'reaction/provider/provider.dart';
 export 'reaction/provider/reaction_animation_widget_state_provider.dart';
 export 'reaction/reaction.dart';
